refactor: エントリーポイントをapp.pyに統合し、ストレージ監視機能を統合
- メインエントリーポイントをapp.pyに変更
- storage_monitor.pyの機能をapp.pyに統合
- ストレージ監視の設定管理とヘルスチェック機能を維持
