+++ backend-db-registration/src/app.py
@@ -23,6 +23,85 @@ app = FastAPI(
 # Initialize webhook file watcher service
 webhook_watcher = WebhookFileWatcherService()
 
+# Initialize storage monitor
+from storage_monitor import StorageMonitor
+storage_monitor = StorageMonitor()
+
+# Add storage monitor health check endpoints
+@app.get("/health/storage-monitor")
+async def storage_monitor_health():
+    """ストレージ監視サービスの詳細ヘルスチェック"""
+    try:
+        # Webhook設定の確認（セキュリティのため詳細情報は制限）
+        webhook_status = {
+            "enabled": storage_monitor.webhook_enabled,
+            "configured": True
+        }
+        
+        # ストレージ接続の確認
+        from storage.storage_client import StorageClient
+        storage_client = StorageClient()
+        try:
+            storage_client.storage_connection_check()
+            storage_status = {"status": "connected", "storage": "minio"}
+        except Exception as e:
+            storage_status = {"status": "error", "error": str(e)}
+        
+        # データベース接続の確認
+        from db.database import engine
+        try:
+            with engine.connect() as conn:
+                from sqlalchemy import text
+                conn.execute(text("SELECT 1"))
+                conn.commit()
+            db_status = {"status": "connected", "database": "postgresql"}
+        except Exception as e:
+            db_status = {"status": "error", "error": str(e)}
+        
+        return {
+            "status": "healthy",
+            "service": "storage-monitor",
+            "webhook": webhook_status,
+            "storage": storage_status,
+            "database": db_status
+        }
+        
+    except Exception as e:
+        logger.error(f"Health check failed: {e}")
+        raise HTTPException(status_code=500, detail=f"Health check failed: {str(e)}")
+
+@app.get("/health/storage-monitor/ready")
+async def storage_monitor_ready():
+    """ストレージ監視サービスの準備完了チェック"""
+    try:
+        # 基本的な接続確認
+        from storage.storage_client import StorageClient
+        from db.database import engine
+        
+        storage_client = StorageClient()
+        
+        try:
+            storage_ok = storage_client.storage_connection_check()
+        except Exception:
+            storage_ok = False
+        try:
+            with engine.connect() as conn:
+                from sqlalchemy import text
+                conn.execute(text("SELECT 1"))
+                conn.commit()
+            db_ok = True
+        except Exception:
+            db_ok = False
+        
+        if storage_ok and db_ok:
+            return {"status": "ready", "service": "storage-monitor"}
+        else:
+            raise HTTPException(status_code=503, detail="Service not ready")
+            
+    except Exception as e:
+        logger.error(f"Readiness check failed: {e}")
+        raise HTTPException(status_code=503, detail=f"Service not ready: storage_ok={storage_ok}, db_ok={db_ok}, error={str(e)}")
+
 
 @app.get("/")
 async def hello():
